refactor: move autolink href helper
