+++ packages/core/src/extensions/link-commands.ts
@@ -2,10 +2,10 @@ import { defineCommands, defineKeymap, isTextSelection, type PlainExtension } fr
 import type { Command, EditorState } from '@prosekit/pm/state'
 import { TextSelection } from '@prosekit/pm/state'
 
+import { getAutolinkHref } from '../lezer/autolink-tld.ts'
 import type { PositionRange } from '../utils/range.ts'
 
 import { getLinkUnitAt, type LinkUnit } from './get-link-unit-at.ts'
-import { getAutolinkHref } from './inline-text-to-mark-chunks.ts'
 import { trimRange } from './inline-toggle.ts'
 
 export interface LinkAttrs {
